requirements.txt
pillow
layoutparser
torch
torchvision
torchaudio
reportlab
pypdf
transformers
sentencepiece
pdf2image
sacremoses
google-generativeai
